Fix: 설계 변경 으로 인한 오류 수정

+++ src/main/java/com/back/splitmeet/src/board/dto/PostCreateReq.java
@@ -23,6 +23,6 @@ public class PostCreateReq {
 	private String localMoneyDescription;
 	private String localAddress;
 	//cobuy
-	private Integer localMoney;
+	private Long localMoney;
 	private String localDescription;
 }
